+++ NEWS.md
@@ -1,3 +1,7 @@
+# solartime 0.0.3
+- add function setLocalTimeZone
+- convert internally to local time zone in computeSunPosition 
+
 # solartime 0.0.2
 - add rmarkdown to suggest dependencies
 - change default doy in computeSolarToLocalTimeDifference from numeric(0) to NA